Correct submission command in multipatch gallery example
The old command contained zero-width spaces to help with line breaking.
Unfortunately, these were preserved by copy and paste, leading to a
command which would not run.  In this commit, change to using CSS nowrap
around the "--parfile .../.par" part to avoid the unsightly break after
"--".  This looks good with Safari and Firefox on Mac OS, and Safari on
iOS, and copy and paste works with Mac OS and Terminal.