remove link from "many different researchers" to maintainers
the text made it looks as if the maintainers had developed all code
in the ET which is not true and never was a claim we wanted to make.
Originally the link pointed to credits.php which contained a list of
authors of commits. This list was static so constantly out of date.
Better to have none instead.